requirements.txt
setuptools~=50.3.2
argparse~=1.4.0
PyYAML~=5.3.1
pykeepass~=4.0.1
SecretStorage~=3.3.1
Flask~=2.0.2
flask-restful~=0.3.9
deepmerge~=0.3.0
apispec~=5.2.1
